refactor response handling
