+++ src/Illuminate/Http/Resources/JsonResourceResponse.php
@@ -28,7 +28,7 @@ class JsonResourceResponse extends ResourceResponse
 
         return $this->build($request, response()->json(
             array_merge_recursive($this->wrap($this->resource->toJson($request)), $this->with),
-            $this->calculateStatus(), $this->headers
+            $this->calculateStatus(), $this->resource->headers
         ));
     }
